Refactor: 게시글 이미지 삭제

+++ src/posts/post.service.ts
@@ -1,5 +1,5 @@
 import { PostsViewDto } from './dto/addPostsView.dto';
-import { BadRequestException, ForbiddenException, Injectable, InternalServerErrorException, NotFoundException } from '@nestjs/common';
+import { BadRequestException, ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from 'src/users/user.entity';
 import { CreatePostDto } from './dto/createPost.dto';
@@ -17,16 +17,17 @@ import { PostsLikeRecord } from './postsLikeRecord.entity';
 import { PostsLikeDto } from './dto/addPostsLike.dto';
 import { PostsViewRecord } from './postsViewRecord.entity';
 import { v1 as uuid } from 'uuid';
-import { createWriteStream } from 'fs';
 import { getRepository } from 'typeorm';
 import { Location } from 'src/users/location.entity';
 import { UserService } from 'src/users/user.service';
 import * as config from 'config';
 import * as AWS from 'aws-sdk';
 import { FileUpload } from 'src/users/models/fileUpload.model';
+import { PostImage } from './postImage.entity';
 
 const s3Config: any = config.get('S3');
 const AWS_S3_BUCKET_NAME = s3Config.AWS_S3_BUCKET_NAME;
+const s3 = new AWS.S3();
 AWS.config.update({
   region: s3Config.AWS_S3_REGION,
   credentials: {
@@ -42,6 +43,55 @@ export class PostService {
     private postRepository: PostRepository,
     private readonly userService: UserService,
   ) {}
+
+  async imagesUploadToS3(insertId: number, images: Promise<FileUpload>[]) {
+    /**
+     * S3에서 게시글 이미지 저장
+     *
+     * @author 허정연(golgol22)
+     * @param {insertId, images} 이미지 저장할 게시글 ID, 업로드할 이미지
+     */
+    for (const image of images) {
+      const { encoding, mimetype, createReadStream } = await image;
+      try {
+        const newFileName = uuid();
+        await s3
+          .putObject({
+            Key: `${newFileName}.png`,
+            Body: createReadStream(),
+            Bucket: `${AWS_S3_BUCKET_NAME}/post`,
+            ContentEncoding: encoding,
+            ContentType: mimetype,
+            ContentLength: createReadStream().readableLength,
+          })
+          .promise();
+        await this.postRepository.addPostImagePath(insertId, `${newFileName}.png`);
+      } catch (err) {
+        console.error(err);
+      }
+    }
+  }
+
+  async imageDeleteFromS3(postImages: PostImage[]) {
+    /**
+     * S3에서 게시글 이미지 삭제
+     *
+     * @author 허정연(golgol22)
+     * @param {postImages} 삭제할 게시글 이미지
+     */
+    for (const file of postImages) {
+      try {
+        await s3
+          .deleteObject({
+            Key: file.imagePath,
+            Bucket: `${AWS_S3_BUCKET_NAME}/post`,
+          })
+          .promise();
+      } catch (err) {
+        console.error(err);
+      }
+    }
+  }
 
   async createPost(user: User, createPostDto: CreatePostDto): Promise<Post> {
     /**
@@ -66,34 +116,6 @@ export class PostService {
     return await this.getPostById(insertId);
   }
 
-  async imagesUploadToS3(insertId: number, images: Promise<FileUpload>[]) {
-    /**
-     * 게시글 이미지 저장
-     *
-     * @author 허정연(golgol22)
-     * @param {insertId, images} 이미지 저장할 게시글 ID, 업로드할 이미지
-     * @throws {InternalServerErrorException} 이미지 저장실패할 때 예외처리
-     */
-    for (const image of images) {
-      const { encoding, mimetype, createReadStream } = await image;
-      try {
-        await new AWS.S3()
-          .putObject({
-            Key: `${uuid()}.png`,
-            Body: createReadStream(),
-            Bucket: `${AWS_S3_BUCKET_NAME}/post`,
-            ContentEncoding: encoding,
-            ContentType: mimetype,
-            ContentLength: createReadStream().readableLength,
-          })
-          .promise();
-        await this.postRepository.addPostImagePath(insertId, `${AWS_S3_BUCKET_NAME}/post/${uuid()}.png`);
-      } catch (err) {
-        console.error(err);
-      }
-    }
-  }
-
   async updatePost(user: User, postId: number, updatePostDto: UpdatePostDto): Promise<Post> {
     /**
      * 게시글 수정
@@ -102,21 +124,21 @@ export class PostService {
      * @param {user, postId, title, content, category, price, isOfferedPrice, townRange, images}
      *        로그인한 유저, 게시글 ID, 제목, 내용, 카테고리, 가격, 가격제안받기여부, 동네범위, 이미지
      * @return {Post} 게시글 반환
-     * @throws {ForbiddenException} 권한없는 사용자의 수정 요청 예외처리
      * @throws {NotFoundException} 해당 게시글이 없을 때 예외 처리
-     * @throws {InternalServerErrorException} 이미지 저장실패할 때 예외처리
+     * @throws {ForbiddenException} 권한없는 사용자의 수정 요청 예외처리
      */
     const post = await this.getPostById(postId);
-    if (JSON.stringify(post.user) !== JSON.stringify(user)) {
-      throw new ForbiddenException(`본인이 작성한 게시글만 수정할 수 있습니다.`);
-    }
     if (!post) {
       throw new NotFoundException(`postId가 ${postId}인 것을 찾을 수 없습니다.`);
     }
+    if (JSON.stringify(post.user) !== JSON.stringify(user)) {
+      throw new ForbiddenException(`본인이 작성한 게시글만 수정할 수 있습니다.`);
+    }
     await this.postRepository.updatePost(postId, updatePostDto);
-    await this.postRepository.deletePostImagePath(postId);
     const { images } = updatePostDto;
     if (images) {
+      await this.imageDeleteFromS3(post.postImages);
+      await this.postRepository.deletePostImagePath(postId);
       await this.imagesUploadToS3(postId, images);
     }
     return await this.getPostById(postId);
@@ -125,22 +147,23 @@ export class PostService {
   async deletePost(user: User, postId: number): Promise<string> {
     /**
      * 게시글 삭제
-     * 게시글 삭제시 DB에 저장된 게시글 이미지 자동 삭제 (Cascade)
+     * 게시글 삭제시 DB에 저장된 게시글 이미지 경로 자동 삭제 (Cascade)
      *
      * @author 허정연(golgol22)
      * @param {user, postId} 로그인한 유저, 게시글 ID
      * @return {string} 삭제되었습니다.
-     * @throws {ForbiddenException} 권한없는 사용자의 삭제 요청 예외처리
      * @throws {NotFoundException} 해당 게시글이 없을 때 예외 처리
+     * @throws {ForbiddenException} 권한없는 사용자의 삭제 요청 예외처리
      */
     const post = await this.getPostById(postId);
+    if (!post) {
+      throw new NotFoundException(`postId가 ${postId}인 것을 찾을 수 없습니다.`);
+    }
     if (JSON.stringify(post.user) !== JSON.stringify(user)) {
       throw new ForbiddenException(`본인이 작성한 게시글만 삭제할 수 있습니다.`);
     }
-    const result = await this.postRepository.delete(postId);
-    if (result.affected === 0) {
-      throw new NotFoundException(`postId가 ${postId}인 것을 찾을 수 없습니다.`);
-    }
+    await this.imageDeleteFromS3(post.postImages);
+    await this.postRepository.delete(postId);
     return '삭제되었습니다.';
   }
